+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- Add badges to the video buttons to show how many videos in each category.
+
 ### Changed
 - Improve responsiveness on startup.